pyproject.toml
[tool.poetry]
name = "word-sandwich"
version = "0.1.0"
description = ""
authors = ["Jake Coble <jakecoble@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
beautifulsoup4 = "^4.9.1"
python-slugify = "^4.0.1"
flask = "^1.1.2"
cssutils = "^1.0.2"
pipe = "^1.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
